ExpensePaymentSystem/ExpensePaymentSystem.UnitTests/Application/PaymentOperations/Queries/GetPaymentById/GetPaymentByIdQueryHandlerTests.cs
using AutoMapper;
using ExpensePaymentSystem.Base.Response;
using ExpensePaymentSystem.Business.Queries.ExpenseQueries.GetExpenseById;
using ExpensePaymentSystem.Business.Queries.PaymentQueries.GetPaymentById;
using ExpensePaymentSystem.Data.DbContext;
using ExpensePaymentSystem.Data.Entity;
using ExpensePaymentSystem.Schema;
using ExpensePaymentSystem.UnitTests.TestsSetup;
using FluentAssertions;
using MediatR;
using Microsoft.EntityFrameworkCore;

namespace ExpensePaymentSystem.UnitTests.Application.PaymentOperations.Queries.GetPaymentById;

public class GetPaymentByIdQueryHandlerTests : IClassFixture<CommonTestFixture>
{
    private readonly ExpensePaymentSystemDbContext _dbContext;
    private readonly IMapper _mapper;

    public GetPaymentByIdQueryHandlerTests(CommonTestFixture fixture)
    {
        _dbContext = fixture.Context;
        _mapper = fixture.Mapper;
    }

    [Fact]
    public async Task Handle_WhenExpenseExists_ShouldReturnPayment()
    {
        // Arrange
        var handler = new GetPaymentByIdQueryHandler(_dbContext, _mapper);
        var query = new GetPaymentByIdQuery(1);
        
        // Act
        var result = await handler.Handle(query, CancellationToken.None);

        // Assert
        result.Should().NotBeNull();
        result.Response.Should().NotBeNull();
        result.Response.PaymentId.Should().Be(1);
    }
}
